feat: implement migration scripts for tags and organizations
